+++ src/app.module.ts
@@ -3,7 +3,6 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
 import { OpenaiClientModule } from './openai-client/openai-client.module';
-import { ChatsModule } from './chats/chats.module';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ChatbotsModule } from './chatbots/chatbots.module';
 import * as process from 'process';
@@ -13,7 +12,6 @@ import * as process from 'process';
     ConfigModule.forRoot({ isGlobal: true }),
     MongooseModule.forRoot(process.env.MONGO_DB_URL),
     OpenaiClientModule,
-    ChatsModule,
     ChatbotsModule,
   ],
   controllers: [AppController],
