--- src/chats/chats.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ChatsService } from './chats.service';
-
-describe('ChatsService', () => {
-  let service: ChatsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ChatsService],
-    }).compile();
-
-    service = module.get<ChatsService>(ChatsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
